backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
accelerate=1.9.0=pyhe01879c_0
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
aiohttp=3.11.10=py312h80987f9_0
aiosignal=1.4.0=pyhd8ed1ab_0
annotated-types=0.7.0=pyhd8ed1ab_1
anyio=4.10.0=pypi_0
aom=3.6.0=h313beb8_0
arrow-cpp=13.0.0=hce30654_42_cpu
async-timeout=5.0.1=py312hca03da5_0
attrs=25.3.0=pyh71513ae_0
aws-c-auth=0.7.22=hec39e38_2
aws-c-cal=0.6.14=h5db4892_1
aws-c-common=0.9.19=h99b78c6_0
aws-c-compression=0.2.18=h5db4892_6
aws-c-event-stream=0.4.2=h5eab607_12
aws-c-http=0.8.1=had10953_17
aws-c-io=0.14.8=hb5a7b21_5
aws-c-mqtt=0.10.4=h78534b8_4
aws-c-s3=0.5.9=h1755d02_3
aws-c-sdkutils=0.1.16=h5db4892_2
aws-checksums=0.1.18=h5db4892_6
aws-crt-cpp=0.26.9=h03bff2b_0
aws-sdk-cpp=1.11.329=hb37a6d0_3
blas=2.131=openblas
blas-devel=3.9.0=31_h11c0a38_openblas
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.4.2=py312ha86b861_0
brotli=1.1.0=pypi_0
brotlicffi=1.0.9.2=py312h313beb8_1
bzip2=1.0.8=h80987f9_6
c-ares=1.34.5=h5505292_0
ca-certificates=2025.8.3=hbd8a1cb_0
cairo=1.16.0=hd886ae0_6
certifi=2025.8.3=pyhd8ed1ab_0
cffi=1.17.1=py312h3eb5a62_1
charset-normalizer=3.4.2=pyhd8ed1ab_0
click=8.2.1=pyh707e725_0
colorama=0.4.6=py312hca03da5_0
cyrus-sasl=2.1.27=h59bf3fd_6
datasets=2.19.1=py312hca03da5_0
dav1d=1.2.1=h80987f9_0
deprecated=1.2.13=py312hca03da5_0
diffusers=0.34.0=pyhd8ed1ab_0
dill=0.3.8=py312hca03da5_0
dnspython=2.7.0=pyhff2d567_1
eigen=3.4.0=h1995070_0
email-validator=2.2.0=pyhd8ed1ab_1
email_validator=2.2.0=hd8ed1ab_1
exceptiongroup=1.3.0=pyhd8ed1ab_0
expat=2.7.0=h286801f_0
fastapi=0.116.1=h26c32bb_1
fastapi-cli=0.0.8=pyhd8ed1ab_0
fastapi-core=0.116.1=pyhe01879c_1
ffmpeg=6.1.1=h7693a39_3
ffmpy=0.6.1=pypi_0
filelock=3.18.0=pyhd8ed1ab_0
fontconfig=2.14.1=h6402c1e_3
freetype=2.13.3=h47d26ad_0
fribidi=1.0.10=h1a28f6b_0
frozenlist=1.5.0=py312h80987f9_0
fsspec=2024.3.1=py312hca03da5_0
gdk-pixbuf=2.42.10=h9bcf4fe_0
gettext=0.21.0=hbdbcc25_2
gflags=2.2.2=h313beb8_1
giflib=5.2.2=h80987f9_0
glib=2.84.2=hc880cf1_0
glib-tools=2.84.2=hc880cf1_0
glog=0.7.1=heb240a5_0
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py312h5c1b81f_0
gradio=5.41.1=pypi_0
gradio-client=1.11.0=pypi_0
graphite2=1.3.14=hc377ac9_1
greenlet=3.1.1=py312hde4cb15_0
groovy=0.1.2=pypi_0
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
h11=0.16.0=pyhd8ed1ab_0
h2=4.2.0=pyhd8ed1ab_0
harfbuzz=10.2.0=he637ebf_1
hdf5=1.14.5=hd77251f_2
hf-xet=1.1.7=pypi_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pyh29332c3_0
httptools=0.6.4=py312hea69d52_0
httpx=0.28.1=pyhd8ed1ab_0
huggingface-hub=0.34.3=pypi_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.1=h313beb8_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.7.0=pyhe01879c_1
jinja2=3.1.6=py312hca03da5_0
jpeg=9e=h80987f9_3
krb5=1.20.1=h69eda48_0
lame=3.100=h1a28f6b_0
lcms2=2.16=he93ba84_0
leptonica=1.82.0=h18c3764_2
lerc=4.0.0=h313beb8_0
libabseil=20240116.2=cxx17_h313beb8_0
libarchive=3.7.7=h8f13d7a_0
libarrow=13.0.0=hc445d41_42_cpu
libavif=1.1.1=h80987f9_0
libblas=3.9.0=31_h10e41b3_openblas
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.1.0=h5505292_3
libbrotlidec=1.1.0=h5505292_3
libbrotlienc=1.1.0=h5505292_3
libcblas=3.9.0=31_hb3479ef_openblas
libclang=14.0.6=default_h5dc8d65_1
libclang13=14.0.6=default_hc7183e1_1
libcrc32c=1.1.2=hc377ac9_0
libcurl=8.12.1=hde089ae_0
libcxx=17.0.6=he5c5206_4
libdeflate=1.22=h80987f9_0
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libexpat=2.7.0=h286801f_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libglib=2.84.2=hdc2269c_0
libgoogle-cloud=2.24.0=hfe08963_0
libgoogle-cloud-storage=2.24.0=h3fa5b87_0
libgrpc=1.62.2=h9c18a4f_0
libiconv=1.16=h80987f9_3
libjpeg-turbo=2.0.0=h1a28f6b_0
liblapack=3.9.0=31_hc9a63f6_openblas
liblapacke=3.9.0=31_hbb7bcf8_openblas
libllvm14=14.0.6=hd1a9a77_4
libnghttp2=1.57.0=h62f6fdd_0
libntlm=1.8=h5505292_0
libogg=1.3.5=h1a28f6b_1
libopenblas=0.3.29=hea593b9_0
libopus=1.3.1=h80987f9_1
libpng=1.6.39=h80987f9_0
libpq=17.4=h02f6b3c_0
libprotobuf=4.25.3=h514c7bf_0
libre2-11=2024.02.01=h12402fc_0
librsvg=2.56.3=h38920c8_2
libsentencepiece=0.2.0=h99b25e3_3
libsqlite=3.46.0=hfb93653_0
libssh2=1.11.1=h3e2b118_0
libtheora=1.1.1=h1a28f6b_3
libthrift=0.19.0=h026a170_1
libtiff=4.5.1=hc9ead59_1
libtorch=2.5.1=gpu_mps_h82d5d13_202
libutf8proc=2.8.0=hc098a78_1
libuv=1.48.0=h80987f9_0
libvorbis=1.3.7=h1a28f6b_0
libvpx=1.13.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_1
libxml2=2.13.8=h0b34f26_0
libzlib=1.2.13=hfb2fe0b_6
llvm-openmp=17.0.6=h2f17746_0
lz4-c=1.9.4=h313beb8_1
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py312h80987f9_0
mdurl=0.1.2=pyhd8ed1ab_1
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py312hca03da5_0
multidict=6.1.0=py312h80987f9_0
multiprocess=0.70.15=py312hca03da5_0
mysql=8.4.0=h3a6587f_1
ncurses=6.4=h313beb8_0
networkx=3.4.2=py312hca03da5_0
nomkl=3.0=0
numexpr=2.11.0=py312h8c72970_0
numpy=1.26.4=py312h7f4fdc5_0
numpy-base=1.26.4=py312he047099_0
openblas=0.3.29=hca03da5_0
openblas-devel=0.3.29=hea593b9_0
opencv=4.10.0=py312h6113af1_2
openh264=2.1.1=h98b2900_0
openjpeg=2.5.0=hbc2ba62_2
openldap=2.6.5=h7c5f683_0
openssl=3.5.2=he92f556_0
opentelemetry-api=1.35.0=pyhd8ed1ab_0
orc=2.0.1=h937ddfc_0
orjson=3.11.1=pypi_0
packaging=25.0=pyh29332c3_1
pandas=2.2.3=py312hcf29cfe_0
pango=1.50.7=hf2328af_2
pcre2=10.42=hb066dcc_1
pillow=11.1.0=py312h84e58ab_0
pip=25.1=pyhc872135_2
pixman=0.40.0=h1a28f6b_0
propcache=0.3.1=py312h80987f9_0
protobuf=4.25.3=py312h1a06a43_1
psutil=5.9.0=py312h80987f9_1
pyarrow=13.0.0=py312he950b47_42_cpu
pycparser=2.22=pyh29332c3_1
pydantic=2.11.7=pyh3cfb1c2_0
pydantic-core=2.33.2=py312hd3c0895_0
pydub=0.25.1=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.2=hdf0ec26_0_cpython
python-dateutil=2.9.0post0=py312hca03da5_2
python-dotenv=1.1.1=pyhe01879c_0
python-multipart=0.0.20=pyhff2d567_0
python-tzdata=2025.2=pyhd3eb1b0_0
python-xxhash=3.5.0=py312h80987f9_0
python_abi=3.12=7_cp312
pytorch=2.5.1=gpu_mps_py312h80af30b_202
pytz=2025.2=py312hca03da5_0
pyyaml=6.0.2=py312h80987f9_0
qt-main=5.15.2=h0917680_12
re2=2024.02.01=h48ca7d4_0
readline=8.2=h1a28f6b_0
regex=2024.11.6=py312h80987f9_0
requests=2.32.4=py312hca03da5_0
rich=14.1.0=pyhe01879c_0
rich-toolkit=0.15.0=pyhe01879c_0
ruff=0.12.7=pypi_0
safehttpx=0.1.6=pypi_0
safetensors=0.5.3=py312hcd83bfe_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.2.0=hca03da5_3
sentencepiece-python=0.2.0=py312h65de07c_3
sentencepiece-spm=0.2.0=h99b25e3_3
setuptools=78.1.1=py312hca03da5_0
shellingham=1.5.4=pyhd8ed1ab_1
six=1.17.0=py312hca03da5_0
sleef=3.5.1=h80987f9_2
snappy=1.2.1=h313beb8_0
sniffio=1.3.1=pyhd8ed1ab_1
sqlalchemy=2.0.43=py312h163523d_0
sqlite=3.45.3=h80987f9_0
starlette=0.47.2=pyh82d4cca_0
sympy=1.13.3=py312hca03da5_1
tesseract=5.2.0=h313beb8_2
tk=8.6.14=h6ba3021_1
tokenizers=0.21.0=py312he2d9c3e_0
tomlkit=0.13.3=pypi_0
torchaudio=2.5.1=py312hd77ebd4_0
torchvision=0.20.1=mps_py312h620d032_102
tqdm=4.67.1=py312h989b03a_0
transformers=4.53.3=pyhd8ed1ab_0
typer=0.16.0=pyh167b9f4_0
typer-slim=0.16.0=pyhe01879c_0
typer-slim-standard=0.16.0=hf964461_0
typing-extensions=4.14.1=h4440ef1_0
typing-inspection=0.4.1=pyhd8ed1ab_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=py312hca03da5_0
utf8proc=2.6.1=h80987f9_1
uvicorn=0.35.0=pyh31011fe_0
uvicorn-standard=0.35.0=h31011fe_0
uvloop=0.20.0=py312h024a12e_0
watchfiles=1.1.0=py312hcd83bfe_0
websockets=15.0.1=py312h290adc7_1
wheel=0.45.1=py312hca03da5_0
wrapt=1.17.0=py312h80987f9_0
xxhash=0.8.0=h1a28f6b_3
xz=5.6.4=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.18.0=py312h80987f9_0
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.2.13=hfb2fe0b_6
zstandard=0.23.0=py312h1a4646a_1
zstd=1.5.6=hfb09047_0
